pyproject.toml
[project]
name = "lecture"
version = "0.1.0"
description = ""
authors = [
    {name = "Ihor Melnyk"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "celery>=5.4.0",
    "Django>=5.1.4",
    "django-cors-headers>=4.6.0",
    "django-crispy-forms>=2.3",
    "django-environ>=0.11.2",
    "django-utils-six>=2.0",
    "gunicorn>=23.0.0",
    "Pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "pyyaml",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "charset-normalizer>=3.4.1,<4.0.0",
    "urllib3>=2.3.0,<3.0.0",
    "poetry>=2.0.0,<3.0.0",
    "black (>=24.10.0,<25.0.0)",
    "ruff (>=0.9.1,<0.10.0)",
    "pytz (>=2024.2,<2025.0)",
    "daphne (>=4.2.1,<5.0.0)",
    "channels (>=4.2.2,<5.0.0)",
    "channels-redis (>=4.3.0,<5.0.0)",
    "aiohttp (>=3.12.14,<4.0.0)",
    "mutagen (>=1.47.0,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
